requirements.txt
paho-mqtt==1.6.1
pyyaml==6.0.1
loguru==0.7.2
